Back-end/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
inflection==0.5.1
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
sqlparse==0.4.4
stripe==8.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
